Move Event form stuff to a Trait
